# Conflicts:
#	src/main/java/MSACHAT/backend/repository/PostRepository.java
#	src/main/java/MSACHAT/backend/service/impl/PostServiceImpl.java

+++ src/main/java/MSACHAT/backend/controller/PostController.java
@@ -8,19 +8,15 @@ import MSACHAT.backend.service.CommentService;
 import MSACHAT.backend.service.PostService;
 import MSACHAT.backend.entity.PostEntity;
 import MSACHAT.backend.mapper.Mapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-
-import static java.lang.System.currentTimeMillis;
 
 @RestController
 @RequestMapping("/post")
@@ -44,18 +40,19 @@ public class PostController {
     }
 
     @PostMapping("/add")
-    public ResponseEntity<String> addPost(@RequestBody PostDto postDto, @RequestHeader("Authorization") String bearerToken
+    public ResponseEntity<Object> addPost(@RequestBody PostDto postDto, @RequestHeader("Authorization") String bearerToken
     ) {
-        if (postDto.getTitle() != null && postDto.getContent() != null && postDto.getImage() != null) {
-            Integer userId = authService.getUserIdFromToken(bearerToken);
-            PostEntity savedPostEntity = postService.addPost(userId, postDto.getTitle(), postDto.getContent());
-            for (String image : postDto.getImage()) {
-                postService.addImage(savedPostEntity, image);
-            }
-
-            return new ResponseEntity<>("success", HttpStatus.OK);
+        if (postDto.getTitle() == null && postDto.getContent() == null && postDto.getImage() == null) {
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>("error", HttpStatus.BAD_REQUEST);
+        Integer userId = authService.getUserIdFromToken(bearerToken);
+        PostEntity savedPostEntity = postService.addPost(userId, postDto.getTitle(), postDto.getContent());
+        for (String image : postDto.getImage()) {
+            postService.addImage(savedPostEntity, image);
+        }
+
+        return new ResponseEntity<>("success: created", HttpStatus.CREATED);
     }
 
 
@@ -102,29 +99,83 @@ public class PostController {
 
 
     @GetMapping("/getbypagenumandpagesize")
-    public List<PostEntity> getPosts(
+    public ResponseEntity<Object> getPosts(
             @RequestHeader("Authorization") String bearerToken,
             @RequestBody PageNumDto pageNumDto) {
-        return postService.findPostsByPageNum(authService.getUserIdFromToken(bearerToken), pageNumDto.getPageNum(),pageNumDto.getPageSize());
+        if (pageNumDto.getPageSize() == null || pageNumDto.getPageNum() == null) {//RequestBody Info Insufficient 10001 Error
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
+        }
+        List<PostEntity> posts = postService.findPostsByPageNum(authService.getUserIdFromToken(bearerToken), pageNumDto.getPageNum(), pageNumDto.getPageSize());
+        return new ResponseEntity<>(posts, HttpStatus.OK);
+    }
+
+    //For API Test
+    @GetMapping("/getbypagenumandpagesize/test")
+    public ResponseEntity<Object> getPostsTest(
+            @RequestBody PageNumDto pageNumDto) {
+        if (pageNumDto.getPageSize() == null || pageNumDto.getPageNum() == null) {//RequestBody Info Insufficient 10001 Error
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
+        }
+        List<PostEntity> posts = postService.findPostsByPageNum(1, pageNumDto.getPageNum(), pageNumDto.getPageSize());
+        return new ResponseEntity<>(posts, HttpStatus.OK);
     }
 
     @PatchMapping("/{id}/like")
-    public void likePost(
+    public ResponseEntity<Object> likePost(
             @PathVariable("id") Integer postId,
             @RequestBody IsLikedDto isLikedDto,
-            @RequestHeader String token) {
+            @RequestHeader String token
+    ) {
+        if (isLikedDto.getIsLiked() == null) {//RequestBody Info Insufficient 10001 Error
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
+        }
+        if (postService.findPostById(postId) == null) {
+            ErrorDto err = new ErrorDto("Post No Longer Exists.", 10002);
+            return new ResponseEntity<>(err, HttpStatus.NOT_FOUND);
+        }
         boolean isLiked = isLikedDto.getIsLiked();
         if (isLiked) {
             postService.unlikePost(postId, authService.getUserIdFromToken(token));
         } else {
             postService.likePost(postId, authService.getUserIdFromToken(token));
         }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
+    //For API Test
+    @PatchMapping("/{id}/like/test")
+    public ResponseEntity<Object> likePostTest(
+            @PathVariable("id") Integer postId,
+            @RequestBody IsLikedDto isLikedDto
+    ) {
+        if (isLikedDto.getIsLiked() == null) {//RequestBody Info Insufficient 10001 Error
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
+        }
+        if (postService.findPostById(postId) == null) {
+            ErrorDto err = new ErrorDto("Post No Longer Exists.", 10002);
+            return new ResponseEntity<>(err, HttpStatus.NOT_FOUND);
+        }
+        boolean isLiked = isLikedDto.getIsLiked();
+        if (isLiked) {
+            postService.unlikePost(postId, 1);
+        } else {
+            postService.likePost(postId, 1);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
     @DeleteMapping("/{id}/delete")
-    public void deletePost(
+    public ResponseEntity<Object> deletePost(
             @PathVariable("id") Integer postId) {
+        if (postService.findPostById(postId) == null) {
+            ErrorDto err = new ErrorDto("Post No Longer Exists.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.NOT_FOUND);
+        }
         postService.deletePost(postId);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/test")
@@ -133,48 +184,57 @@ public class PostController {
     }
 
     @GetMapping("/{id}/get")
-    public ResponseEntity<PostReturnDto> getPostById(
+    public ResponseEntity<Object> getPostById(
             @PathVariable("id") Integer postId,
             @RequestHeader("Authorization") String bearerToken) {
         String token = authService.getTokenFromHeader(bearerToken);
         Integer userId = authService.getUserIdFromToken(token);
 
-        PostEntity post = postService.findPostById(postId, userId);
+        PostEntity post = postService.findPostByIdAndUserId(postId, userId);
         if (post == null) {
-            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            ErrorDto err = new ErrorDto("Post No Longer Exists.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.NOT_FOUND);
         }
 
         List<String> imageList = post.getImages().stream().map(ImageEntity::getImageUrl).toList();
-        PostReturnDto postReturn = new PostReturnDto(post.getId(),post.getUserId(), post.getTitle(), post.getContent(), imageList,post.getTimeStamp(),post.getLikeCount(),post.getCommentCount(),post.isLiked());
+        PostReturnDto postReturn = new PostReturnDto(post.getId(), post.getUserName(), post.getTitle(), post.getContent(), imageList, post.getTimeStamp(), post.getLikeCount(), post.getCommentCount(), post.isLiked());
 
         return new ResponseEntity<>(postReturn, HttpStatus.OK);
     }
 
     //测试使用
     @GetMapping("/{id}/get/test")
-    public ResponseEntity<PostReturnDto> getPostByIdTest(@PathVariable("id") Integer postId) {
+    public ResponseEntity<Object> getPostByIdTest(@PathVariable("id") Integer postId) {
 
         Integer userId = 1;
 
 
-        PostEntity post = postService.findPostById(postId, userId);
+        PostEntity post = postService.findPostByIdAndUserId(postId, userId);
         if (post == null) {
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
 
         List<String> imageList = post.getImages().stream().map(ImageEntity::getImageUrl).toList();
-        PostReturnDto postReturn = new PostReturnDto(post.getId(),post.getUserId(), post.getTitle(), post.getContent(), imageList,post.getTimeStamp(),post.getLikeCount(),post.getCommentCount(),post.isLiked());
+        PostReturnDto postReturn = new PostReturnDto(post.getId(), post.getUserName(), post.getTitle(), post.getContent(), imageList, post.getTimeStamp(), post.getLikeCount(), post.getCommentCount(), post.isLiked());
 
         return new ResponseEntity<>(postReturn, HttpStatus.OK);
     }
 
 
     @PutMapping("/{id}/comment")
-    public ResponseEntity<String> addComment(
+    public ResponseEntity<Object> addComment(
             @RequestBody CommentInfoDto commentInfo,
             @PathVariable("id") Integer postId,
             @RequestHeader("Authorization") String bearerToken
-                                             ) {
+    ) {
+        if (commentInfo.getContent() == null) {
+            ErrorDto err = new ErrorDto("Request body incomplete. Required fields missing.", 10001);
+            return new ResponseEntity<>(err, HttpStatus.BAD_REQUEST);
+        }
+        if (postService.findPostById(postId) == null) {
+            ErrorDto err = new ErrorDto("Post No Longer Exists.", 10002);
+            return new ResponseEntity<>(err, HttpStatus.NOT_FOUND);
+        }
         String token = authService.getTokenFromHeader(bearerToken);
         Integer userId = authService.getUserIdFromToken(token);
         String content = commentInfo.getContent();
